somthing error, datepicker, initiDateGime, etc

+++ lib/home/calendar/popup/function/change/pop_change_widget.dart
@@ -4,6 +4,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../../../database/database.dart';
 import '../../../../home_page.dart';
 import '../../popup_change_page.dart';
+import '../../provider/pop_show_provider.dart';
 import '../../provider/pop_change_provider.dart';
 import 'pop_change_function.dart';
 
@@ -72,41 +73,37 @@ void startDatePickerChangeFunc(
                   Padding(
                     padding: const EdgeInsets.only(right: 10),
                     child: CupertinoButton(
-                        child: const Text(
-                          '完了',
-                          style: TextStyle(
-                            color: Colors.blue,
-                            fontSize: 16,
-                          ),
+                      child: const Text(
+                        '完了',
+                        style: TextStyle(
+                          color: Colors.blue,
+                          fontSize: 16,
                         ),
-                        onPressed: () async {
-                          Navigator.pop(
-                            context,
-                            CupertinoPageRoute(
-                              builder: (context) => const PopChangeWidget(),
-                            ));
-                            DateTime dateTime = ref.watch(scheStartDateChangeShowProvider);
-                            if (ref
-                                    .watch(scheEndDateChangeShowProvider)
-                                    .isBefore(dateTime.subtract(const Duration(hours: 1))) ||
-                                ref
-                                    .watch(scheEndDateChangeShowProvider)
-                                    .isAtSameMomentAs(dateTime.subtract(const Duration(hours: 1)))) {
-
-                              ref.watch(scheEndDateMinimunShowProvider.notifier).state = DateTime(
-                                    dateTime.year,
-                                    dateTime.month,
-                                    dateTime.day,
-                                    dateTime.hour + 1,
-                                    dateTime.minute);
-                              ref.watch(scheEndDateInitialShowProvider.notifier).state = DateTime(
-                                    dateTime.year,
-                                    dateTime.month,
-                                    dateTime.day,
-                                    dateTime.hour + 1,
-                                    dateTime.minute);
-                           }
-                            }),
+                      ),
+                      onPressed: () async {
+                        DateTime dateTime = ref.watch(scheStartDateChangeShowProvider);
+                        if (ref.watch(scheEndDateMinimunShowProvider).isBefore(ref.watch(scheStartDateChangeShowProvider))) {
+                          ref.watch(scheEndDateMinimunShowProvider.notifier).state = DateTime(
+                                  dateTime.year,
+                                  dateTime.month,
+                                  dateTime.day,
+                                  dateTime.hour + 1,
+                                  dateTime.minute);
+                            ref.watch(scheEndDateInitialShowProvider.notifier).state = DateTime(
+                                  dateTime.year,
+                                  dateTime.month,
+                                  dateTime.day,
+                                  dateTime.hour + 1,
+                                  dateTime.minute);
+                        }
+                        Navigator.pop(
+                          context,
+                          CupertinoPageRoute(
+                            builder: (context) => const PopChangeWidget(),
+                          ),
+                        );
+                      }
+                    ),
                   ),
                 ],
               ),
@@ -140,27 +137,42 @@ void endDatePickerChangeFunc(
   showCupertinoModalPopup(
     context: context,
     builder: (BuildContext context) {
-      
       DateTime dateTime = ref.watch(scheStartDateChangeShowProvider);
-      if (ref
-              .watch(scheEndDateChangeShowProvider)
-              .isBefore(dateTime.subtract(const Duration(hours: 1))) ||
-          ref
-              .watch(scheEndDateChangeShowProvider)
-              .isAtSameMomentAs(dateTime.subtract(const Duration(hours: 1)))) {
+      if (ref.watch(switchChangeProvider) == false) {
+        if (ref
+                .watch(scheEndDateChangeShowProvider)
+                .isBefore(dateTime.subtract(const Duration(hours: 1))) ||
+            ref
+                .watch(scheEndDateChangeShowProvider)
+                .isAtSameMomentAs(dateTime.subtract(const Duration(hours: 1)))) {
 
-        ref.watch(scheEndDateMinimunShowProvider.notifier).state = DateTime(
-              dateTime.year,
-              dateTime.month,
-              dateTime.day,
-              dateTime.hour + 1,
-              dateTime.minute);
-        ref.watch(scheEndDateInitialShowProvider.notifier).state = DateTime(
-              dateTime.year,
-              dateTime.month,
-              dateTime.day,
-              dateTime.hour + 1,
-              dateTime.minute);
+          ref.watch(scheEndDateMinimunShowProvider.notifier).state = DateTime(
+                dateTime.year,
+                dateTime.month,
+                dateTime.day,
+                dateTime.hour + 1,
+                dateTime.minute);
+          ref.watch(scheEndDateInitialShowProvider.notifier).state = DateTime(
+                dateTime.year,
+                dateTime.month,
+                dateTime.day,
+                dateTime.hour + 1,
+                dateTime.minute);
+        }
+        if (ref.watch(scheStartDateChangeShowProvider) == ref.watch(scheEndDateChangeShowProvider)) {
+          ref.watch(scheEndDateMinimunShowProvider.notifier).state = DateTime(
+                dateTime.year,
+                dateTime.month,
+                dateTime.day,
+                dateTime.hour + 1,
+                dateTime.minute);
+          ref.watch(scheEndDateInitialShowProvider.notifier).state = DateTime(
+                dateTime.year,
+                dateTime.month,
+                dateTime.day,
+                dateTime.hour + 1,
+                dateTime.minute);
+        }
       }
 
       return Container(
@@ -209,6 +221,10 @@ void endDatePickerChangeFunc(
                       );
                       await ref.read(
                           popSelectedChangeEndDateProvider(scheEndDate).future);
+                      if (ref.watch(scheEndDateChangeShowProvider).isBefore(ref.watch(scheEndDateInitialShowProvider))) {
+                        ref.watch(scheEndDateChangeShowProvider.notifier).state = ref.watch(scheEndDateInitialShowProvider);
+                      }
+                      print('apple: ${ref.watch(scheEndDateChangeShowProvider)} ${ref.watch(scheEndDateInitialShowProvider)}');
                     },
                   ),
                 ),
@@ -238,30 +254,62 @@ void endDatePickerChangeFunc(
 
 void deleteScheduleFunc(BuildContext context, WidgetRef ref) {
   showCupertinoModalPopup(
-      context: context,
-      builder: (BuildContext context) {
-        return CupertinoActionSheet(
-          actions: [
-            CupertinoActionSheetAction(
-              onPressed: () async {
-                Navigator.push(
-                  context,
-                  CupertinoPageRoute(
-                    builder: (context) => const HomeScreen(),
-                  ),
-                );
-                final database = ref.watch(databaseProvider);
-                await database
-                    .deleteSchedule(ref.watch(databaseGetDateProvider));
-                await database.close();
-              },
-              child: const Text('編集を破棄'),
-            ),
-          ],
-          cancelButton: CupertinoActionSheetAction(
-            onPressed: () => Navigator.pop(context),
-            child: const Text('キャンセル'),
+    context: context,
+    builder: (BuildContext context) {
+      return CupertinoActionSheet(
+        actions: [
+          CupertinoActionSheetAction(
+            child: const Text('編集を破棄'),
+            onPressed: () {
+              showCupertinoDialog<void>(
+                context: context,
+                builder: (BuildContext context) {
+                  return CupertinoAlertDialog(
+                    title: const Text('予定の削除'),
+                    content: const Text('本当にこの日の予定を削除しますか？'),
+                    actions: <CupertinoDialogAction>[
+                      CupertinoDialogAction(
+                        onPressed: () {
+                          Navigator.pop(context);
+                        },
+                        isDestructiveAction: true,
+                        child: const Text(
+                          'キャンセル',
+                          style: TextStyle(
+                            color: Colors.blue,
+                          ),
+                        ),
+                      ),
+                      CupertinoDialogAction(
+                        onPressed: () async{
+                          Navigator.push(
+                            context,
+                            CupertinoPageRoute(
+                              builder: (context) => const HomeScreen(),
+                            ),
+                          );
+                          final database = ref.watch(databaseProvider);
+                          await database
+                              .deleteSchedule(ref.watch(databaseGetDateProvider));
+                          ref.invalidate(scheduleFromDatetimeProvider(ref.watch(databaseGetDateProvider)));
+                        },
+                        child: const Text('削除'),
+                      ),
+                    ],
+                  );
+                },
+              );
+            }
           ),
-        );
-      });
+        ],
+        cancelButton: CupertinoActionSheetAction(
+          onPressed: () => Navigator.pop(context),
+          child: const Text('キャンセル'),
+        ),
+      );
+    }
+  );
 }
+
+
+
